return list of questions in search route
