+++ .vscode/launch.json
@@ -1,6 +1,14 @@
 {
     "configurations": [
         {
+            "name": "Python : fichier actif",
+            "type": "python",
+            "request": "launch",
+            "program": "${file}",
+            "console": "integratedTerminal",
+            "justMyCode": true
+        },
+        {
             "name": "Docker: Python - Flask",
             "type": "docker",
             "request": "launch",
