+++ Dockerfile
@@ -22,4 +22,4 @@ RUN adduser -u 5678 --disabled-password --gecos "" appuser && chown -R appuser /
 USER appuser
 
 # During debugging, this entry point will be overridden. For more information, please refer to https://aka.ms/vscode-docker-python-debug
-CMD ["gunicorn", "--bind", "0.0.0.0:3001", "src.main:app"]
+CMD ["gunicorn", "--bind", "0.0.0.0:3001", "main:app"]
