+++ docker-compose.yml
@@ -3,7 +3,7 @@ version: '3.4'
 services:
   backend:
     image: backend
-    command: gunicorn --bind 0.0.0.0:3001 src.main:app --reload
+    command: gunicorn --bind 0.0.0.0:3001 main:app --reload
     build:
       context: .
       dockerfile: ./Dockerfile
