+++ main.py
@@ -1,0 +1,53 @@
+import random
+
+from dotenv import load_dotenv
+from flask import Flask, jsonify, request
+from requests import Response
+
+from src.answers import get_answers
+from src.SO.so_utils import generate_questions_csv, get_questions
+
+QUERY_RESULTS_PATH = "./BD/QueryResults.csv"
+QUESTIONS_PATH = "./BD/questions.csv"
+
+load_dotenv()
+
+app = Flask(__name__)
+generate_questions_csv(
+    input_path=QUERY_RESULTS_PATH,
+    output_path=QUESTIONS_PATH,
+)
+
+
+@app.route("/")
+def home():
+    return "Hello, Flask!"
+
+
+@app.route("/answers/<question_id>", methods=['GET'])
+def answers(question_id: int) -> Response:
+    print(f"GET /answers/{question_id}")
+    answers = get_answers(question_id)
+    return jsonify(answers)
+
+
+@app.route("/search", methods=['GET'])
+def hello_there():
+    query = request.args.get("q", default="", type=str)
+    print(f"GET /search?q={query}")
+
+    questions = get_questions(QUESTIONS_PATH)
+
+    response = {
+        "query": query,
+        "relatedQuestions": [
+            random.choice(questions),
+            random.choice(questions),
+        ]
+    }
+
+    return jsonify(response)
+
+
+if __name__ == "__main__":
+    app.run(host="0.0.0.0", debug=True)
